requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2

# Email validation for Pydantic's EmailStr
email-validator==2.2.0

# Database
sqlalchemy[asyncio]==2.0.35
psycopg2-binary==2.9.9
asyncpg==0.29.0
greenlet==3.1.1

# HTTP Client
httpx==0.27.2

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
argon2-cffi==23.1.0
python-multipart==0.0.12

# Utilities
python-dotenv==1.0.1

